file-search-server-v2.1/requirements-native.txt
# Requirements for Native macOS Document Search App
# Based on PyWebView integration with React/Flask

# Core WebView framework
pywebview>=4.4.0

# HTTP requests for server communication  
requests>=2.31.0

# Flask backend dependencies (from existing backend)
Flask>=2.3.3
Flask-CORS>=4.0.0
ollama>=0.1.7

# Database access
sqlite3  # Built-in with Python

# Optional: Better JSON handling
ujson>=5.8.0

# Development dependencies
pytest>=7.4.0  # For testing
black>=23.7.0  # Code formatting